Seperate fetching mission data logic
